+++ doc/changelog.md
@@ -15,6 +15,11 @@ Sections
 ### Developers
 -->
 
+## [1.2.2] - 2019-06-06
+### Added
+- x cursor in chart lets user go back in time
+- status line at bottom of page
+
 ## [1.2.1] - 2019-05-26
 ### Added
 - Homekit support for Hue lights